30 May Update

+++ database/migrations/2025_05_22_010721_detail_penjualan_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('PenjualanID')->nullable();
             $table->string('ProdukID')->nullable();
             $table->integer('JumlahProduk')->nullable();
-            $table->decimal('Subtotal', 10, 2)->nullable();
+            $table->bigInteger('Subtotal')->nullable();
             $table->foreign('PenjualanID')->references('PenjualanID')->on('penjualan')->onDelete('cascade');
             $table->foreign('ProdukID')->references('ProdukID')->on('produk')->onDelete('cascade');
             $table->timestamps();
